DOIScanner.java
Add logic to extract "date_created" from resource registry.

Add logic to extract "date_created" from resource registry so that
"publicationYear" is derived from the creation date of the resource
and not the current year of when the DOI registration attempt is
made

Add logic to limit DOI-less resources to only those resource that
are active (i.e., not deleted based on the date_deactivated).

Fine tune logic for creating the DataCite metadata object.