DataPackageCitationServlet.java
Remove "url"  GET parameter; add "packageid" GET parameter.

The "url" GET parameter for defining a package identifier was not used
in any dependent calls, so it was removed (sort of).  Add new GET parameter
"packageid" as replacement for the previously removed "url" parameter.

Note that this class is almost identical to MapBrowseServlet.java.