DataPackageManagerClient.java
Modify URL for read data entity to re-encode "%" character.

Because the HttpGet of the Apache HttpClient performs character reference
value decoding by default, literal values that contain a "%" value are decoded
accordingly when they should remain literal.  This kludge replaces
the "%" character with its decimal character reference value "%25" so that the
HttpGet method changes it back to its literal "%" character.