DataPackageManagerResource.java
Refactor createDataPackage method for threaded operation.

Refactor the createDataPackage method so that the heavy lifting operation
of creating a new data package is performed in a separate thread, while the
web service responds with an 202 Accepted response.  Supporting changes
include the creation of the Creator class (the threaded class), the ErrorLogger
class for writing exceptions to a file in the data package metadata directory,
and a readDataPackageError web sevice method to return any error message.