DataManagerClient.java, DataPackageManager.java, DataPackageManagerResource.java
Refactor the readDataPackageEntity processing to return a File object.

Refactor the readDataPackageEntity processing to return a File object
instead of the raw entity bytes.  This resulted in the creation of two new
methods, getDataEntityFile in both the DataManagerClient and
DataPackageManager classes and making the return object in the
readDataEntity a File object in the DataPackageManagerResource class.