AdvancedSearchServlet.java, AdvancedSearch.java, advancedSearch.jsp, map_functions.js
Add tracked "boundsChangedCount" variable to invoke spatial search.

Add a tracking variable called "boundsChangedCount" that is passed as a hidden
form variable from the advancedSearch.jsp to the AdvancedSearchServlet.java;
a spatial change occurs if the bounds of the map extent change or if the form element
content has changed.